Update signs.html
altered the scripts section and added multiple functions to aid the math portion. there should be a function for a dictionary where the user can input a classmates name to check if the student is a vampire. it should input as "first name" and "last name" then check the signs, and confirm to bring you to the results page.